requirements.txt
aiogram==3.13.1
aiogram_dialog==2.1.0
environs==11.0.0
fluentogram==1.1.6
asyncpg==0.28.0
PyYAML==6.0.1
python-dotenv==1.0.1
aiosqlite==0.19.0
APScheduler==3.10.4
SQLAlchemy==2.0.23
aiolavapy==0.1.2
yookassa-async==0.1.5
YooMoney-async==0.0.6
cryptomus==1.1
aiocryptopay==0.4.5
tinkoff-acquiring==0.1.3
pytz==2023.3